comments apply
